Add sanitize method to applicationController

+++ app/controllers/books_controller.rb
@@ -1,10 +1,12 @@
 class BooksController < ApplicationController
   
   before_action :authenticate_user
+  before_action :book_params_sanitize,{only:[:update]}
+  before_action :book_status_sanitize,{only:[:index]}
+  before_action :book_id_sanitize,{only:[:show,:rental,:putback,:update_form,:update,:delete]}
   
   def index
     book_id_list = []
-    @status = params[:status]
     histories = History.where(status: "rental")
     histories.each do |history|
       book_id_list.push(history.book_id)
@@ -17,15 +19,16 @@ class BooksController < ApplicationController
   end
   
   def show
-    @book = Book.find_by(id: params[:book_id])
+    @book = book_find_by_id(@book_id)
   end
   
   def rental
-    if History.find_by(book_id: params[:book_id],status:"rental")
+    book_find_by_id(@book_id)
+    if History.find_by(book_id: @book_id,status:"rental")
       flash[:notice] = "このほんは、かしだしちゅうです"
-      redirect_to("/books/#{params[:book_id]}")
+      redirect_to("/books/#{@book_id}")
     else
-      history = History.new(user_id: @current_user.id,book_id: params[:book_id],status:"rental")
+      history = History.new(user_id: @current_user.id,book_id: @book_id,status:"rental")
       history.save
       flash[:notice] = "このほんをかりました。かしだしきかんは１しゅうかんです"
       redirect_to("/users/#{@current_user.id}/rental")
@@ -33,7 +36,8 @@ class BooksController < ApplicationController
   end
   
   def putback
-    history = History.find_by(user_id: @current_user.id,book_id: params[:book_id],status:"rental")
+    book_find_by_id(@book_id)
+    history = History.find_by(user_id: @current_user.id,book_id: @book_id,status:"rental")
     if history
       history.status = "putback"
       history.save
@@ -41,7 +45,7 @@ class BooksController < ApplicationController
       redirect_to("/users/#{@current_user.id}/putback")
     else
       flash[:notice] = "このほんは、かりていません"
-      redirect_to("/books/#{params[:book_id]}")
+      redirect_to("/books/#{@book_id}")
     end
   end
   
@@ -69,15 +73,13 @@ class BooksController < ApplicationController
   end
   
   def update_form
-    @book = Book.find_by(id: params[:book_id])
+    @book = Book.find_by(id: @book_id)
     @name = @book.name
     @content = @book.content
   end
   
   def update
-    @name = params[:name]
-    @content = params[:content]
-    @book = Book.find_by(id:params[:book_id])
+    @book = Book.find_by(id: @book_id)
     @book.name = @name
     @book.content = @content
     if @book.save
@@ -94,15 +96,17 @@ class BooksController < ApplicationController
   end
   
   def delete
-    book = Book.find_by(id: params[:book_id])
-    histories = History.where(book_id: params[:book_id])
-    File.delete("public/book_images/#{book.image_url}")
+    book = Book.find_by(id: @book_id)
+    histories = History.where(book_id: [@book_id])
+    if book.image_url
+      File.delete("public/book_images/#{book.image_url}")
+    end
     if book.delete && histories.delete_all
       flash[:notice] = "本の情報を削除しました"
-      redirect_to("/books/putback")
+      redirect_to("/books/putback/index")
     else
       flash[:notice] = "本の情報を削除できませんでした"
-      redirect_to("/books/putback")
+      redirect_to("/books/putback/index")
     end
   end
